Support singularity vncserver

+++ vncserver/start-template-v3.sh
@@ -43,8 +43,9 @@ start_rootless_docker() {
 echo '#!/bin/bash' > ${resource_jobdir}/cancel.sh
 chmod +x ${resource_jobdir}/cancel.sh
 echo "mv ${resource_jobdir}/cancel.sh ${resource_jobdir}/cancel.sh.executed" >> ${resource_jobdir}/cancel.sh
-echo "scancel ${SLURM_JOB_ID}"  >> ${resource_jobdir}/cancel.sh
-
+if ![ -z "${SLURM_JOB_ID}" ]; then
+    echo "scancel ${SLURM_JOB_ID}"  >> ${resource_jobdir}/cancel.sh
+fi
 ###################
 ###################
 
@@ -131,17 +132,36 @@ if [ -z ${service_vnc_exec} ] || ! [ -f "${service_vnc_exec}" ]; then
         exit 1
     fi
     echo "$(date): vncserver is not installed. Using singularity container..."
-    service_vnc_exec="singularity exec --bind /tmp/.X11-unix:/tmp/.X11-unix --bind ${HOME}:${HOME} ${service_vncserver_sif} vncserver"
-    service_vnc_type="TurboVNC"
+    singularity_exec="singularity exec --bind /tmp/.X11-unix:/tmp/.X11-unix --bind ${HOME}:${HOME} ${service_vncserver_sif}"
+    service_vnc_exec="${singularity_exec} vncserver"
+    service_vnc_type="SingularityTurboVNC"
     service_desktop="echo Starting no service desktop on the host"
     mkdir -p /tmp/.X11-unix
     rm -f ~/.vnc/xstartup.turbovnc
 cat >> ~/.vnc/xstartup.turbovnc <<HERE
+#!/bin/sh
 unset SESSION_MANAGER
 unset DBUS_SESSION_BUS_ADDRESS
-startxfce4 &
+HERE
+cat >> ${resource_jobdir}/vncserver.sh <<HERE
+#!/bin/bash
+vncserver ${DISPLAY} -SecurityTypes None
+mkdir -p /run/user/\$(id -u)
+chown "\$(id -u):\$(id -g)" /run/user/\$(id -u)
+export XDG_RUNTIME_DIR=/run/user/\$(id -u)
+export DISPLAY=${DISPLAY}
+export XAUTHORITY="\$HOME/.Xauthority"
+mkdir -p \$HOME/.run
+export XDG_RUNTIME_DIR=\$HOME/.run
+chmod 700 \$HOME/.run
+addr=\$(dbus-daemon --session --fork --print-address)
+export DBUS_SESSION_BUS_ADDRESS="\$addr"
+mkdir -p "\$HOME/.config"
+chmod 700 "\$HOME/.config"
+startxfce4 --replace
 HERE
     chmod +x ~/.vnc/xstartup.turbovnc
+    chmod +x ${resource_jobdir}/vncserver.sh
 fi
 
 if [ -z ${service_vnc_type} ]; then
@@ -218,7 +238,7 @@ HERE
 fi
 
 
-if [[ "${service_vnc_type}" == "TigerVNC" || "${service_vnc_type}" == "TurboVNC" ]]; then
+if [[ "${service_vnc_type}" == "TigerVNC" ]]; then
     #########
     # NoVNC #
     #########
@@ -253,6 +273,7 @@ if [[ "${service_vnc_type}" == "TigerVNC" || "${service_vnc_type}" == "TurboVNC"
     # Start service
     mkdir -p ~/.vnc
     ${service_vnc_exec} -kill ${DISPLAY}
+    echo "${service_vnc_exec} -kill ${DISPLAY}" >> cancel.sh
 
     # To prevent the process from being killed at startime    
     if [ -f "${HOME}/.vnc/xstartup" ]; then
@@ -283,10 +304,8 @@ if [[ "${service_vnc_type}" == "TigerVNC" || "${service_vnc_type}" == "TurboVNC"
         # FIXME: Change ~/.vnc/config
         ${service_vnc_exec} ${DISPLAY} &> ${resource_jobdir}/vncserver.log &
         echo $! > ${resource_jobdir}/vncserver.pid
-    elif [[ "${service_vnc_type}" == "TigerVNC" ]]; then
+    else
         ${service_vnc_exec} ${DISPLAY} -SecurityTypes VncAuth -PasswordFile ${resource_jobdir}/.vncpasswd
-    elif [[ "${service_vnc_type}" == "TurboVNC" ]]; then
-        ${service_vnc_exec} ${DISPLAY} -SecurityTypes None
     fi
 
     rm -f ${resource_jobdir}/service.pid
@@ -316,12 +335,25 @@ if [[ "${service_vnc_type}" == "TigerVNC" || "${service_vnc_type}" == "TurboVNC"
 
     cd ${service_novnc_install_dir}
     
-    echo "Running ./utils/novnc_proxy --vnc ${HOSTNAME}:${displayPort} --listen ${HOSTNAME}:${service_port}"
     ./utils/novnc_proxy --vnc ${HOSTNAME}:${displayPort} --listen ${HOSTNAME}:${service_port} </dev/null &
     echo $! >> ${resource_jobdir}/service.pid
     pid=$(ps -x | grep vnc | grep ${displayPort} | awk '{print $1}')
     echo ${pid} >> ${resource_jobdir}/service.pid
     rm -f ${portFile}
+elif [[ "${service_vnc_type}" == "SingularityTurboVNC" ]]; then
+    # Start service
+    mkdir -p ~/.vnc
+    ${service_vnc_exec} -kill ${DISPLAY}
+    echo "${service_vnc_exec} -kill ${DISPLAY}" >> cancel.sh
+    ${singularity_exec} ${resource_jobdir}/vncserver.sh | tee -a vncserver.out &
+    echo "kill $! # singularity exec" >> cancel.sh
+
+    ./utils/novnc_proxy --vnc ${HOSTNAME}:${displayPort} --listen ${HOSTNAME}:${service_port} </dev/null &
+    echo "kill $! # novnc_proxy" >> cancel.sh
+    pid=$(ps -x | grep vnc | grep ${displayPort} | awk '{print $1}')
+    echo ${pid} >> ${resource_jobdir}/service.pid
+    rm -f ${portFile}
+
 elif [[ "${service_vnc_type}" == "KasmVNC" ]]; then
     ###########
     # KasmVNC #
@@ -338,7 +370,7 @@ elif [[ "${service_vnc_type}" == "KasmVNC" ]]; then
 
 
     ${service_vnc_exec} -kill ${DISPLAY}
-    echo "${service_vnc_exec} -kill ${DISPLAY}" >> cancel.sh.sh
+    echo "${service_vnc_exec} -kill ${DISPLAY}" >> cancel.sh
 
     MAX_RETRIES=5
     RETRY_DELAY=5
